virtusa/requirements.txt
langchain==0.1.14
langchain-openai==0.1.3
langgraph==0.0.29
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.3.8
